resdesign pages to match v3 + PR feedback
