+++ data_explorer/app/pages/dataset.py
@@ -11,9 +11,6 @@ class DatasetExplorerApp(DatasetLoaderApp):
     @staticmethod
     def setup_app_page(dataframe, fields):
         """Build the dataframe explorer table."""
-        st.write("## Dataframe explorer")
-        st.write("In this table, you can explore the dataframe")
-
         image_fields = get_image_fields(fields)
 
         # get the first rows of the dataframe
@@ -63,6 +60,6 @@ class DatasetExplorerApp(DatasetLoaderApp):
 
 
 app = DatasetExplorerApp()
-app.create_common_sidebar()
+app.create_common_interface()
 df, df_fields = app.create_loader_widget()
 app.setup_app_page(df, df_fields)
